pyproject.toml
[project]
name = "gcp-ml-framework"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "kfp>=2.15,<3",
    "google-cloud-aiplatform>=1.136",
    "google-cloud-bigquery>=3.25",
    "google-cloud-storage>=2.18",
    "google-cloud-secret-manager>=2.21",
    "apache-airflow>=2.10",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "typer>=0.15",
    "rich>=13",
    "pyyaml>=6",
    "duckdb>=1.1",
    "pyarrow>=18",
    "pandas>=2.2",
    "apache-airflow-providers-google>=20.0",
]

[project.scripts]
gml = "gcp_ml_framework.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8",
    "pytest-cov",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
